Hookify Time#to_nice_s
